+++ src/__tests__/handler/outline.test.ts
@@ -73,14 +73,20 @@ describe('symbols outline', () => {
 
   describe('configuration', () => {
     it('should follow cursor', async () => {
-      await createBuffer()
+      await createBuffer(`  class myClass {
+  fun1() { }
+  fun2() {}
+}`)
       let curr = await nvim.call('bufnr', ['%'])
       await symbols.showOutline(0)
       let bufnr = await nvim.call('bufnr', ['%'])
       await nvim.command('wincmd p')
       await nvim.command('exe 3')
-      await events.fire('CursorHold', [curr])
+      await events.fire('CursorHold', [curr, [3, 1]])
       await helper.wait(50)
+      await nvim.call('cursor', [1, 1])
+      await events.fire('CursorHold', [curr, [1, 1]])
+      await helper.wait(30)
       let buf = nvim.createBuffer(bufnr)
       let lines = await buf.getLines()
       expect(lines.slice(1)).toEqual([
@@ -95,12 +101,12 @@ describe('symbols outline', () => {
         priority: 10,
         group: 'CocTree'
       })
+      await nvim.command(`bd ${bufnr}`)
+      await events.fire('CursorHold', [curr, [3, 1]])
     })
 
     it('should not follow cursor', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.followCursor': false,
-      })
+      helper.updateConfiguration('outline.followCursor', false,)
       await createBuffer()
       let curr = await nvim.call('bufnr', ['%'])
       await symbols.showOutline(0)
@@ -115,9 +121,7 @@ describe('symbols outline', () => {
     })
 
     it('should keep current window', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.keepWindow': true,
-      })
+      helper.updateConfiguration('outline.keepWindow', true)
       await createBuffer()
       let curr = await nvim.call('bufnr', ['%'])
       await symbols.showOutline()
@@ -126,23 +130,27 @@ describe('symbols outline', () => {
     })
 
     it('should check on buffer switch', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.checkBufferSwitch': true,
-      })
-      await createBuffer()
+      helper.updateConfiguration('outline.checkBufferSwitch', true)
+      let b = await createBuffer()
       await symbols.showOutline(1)
-      await helper.edit('unnamed')
-      await helper.wait(200)
       let buf = await getOutlineBuffer()
+      let bufnr = buf.id
+      await helper.edit('unnamed')
+      await helper.waitValue(async () => {
+        let buf = await getOutlineBuffer()
+        return buf.id > bufnr
+      }, true)
+      buf = await getOutlineBuffer()
       let lines = await buf.lines
       expect(lines[0]).toMatch('Document symbol provider not found')
+      await nvim.command(`bd! ${b.id}`)
+      await helper.wait(10)
+      let loaded = await buf.loaded
+      expect(loaded).toBe(true)
     })
 
     it('should not check on buffer switch', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.checkBufferSwitch': false
-      })
-      await helper.wait(30)
+      helper.updateConfiguration('outline.checkBufferSwitch', false)
       await createBuffer()
       await symbols.showOutline(1)
       await helper.edit('unnamed')
@@ -155,15 +163,27 @@ describe('symbols outline', () => {
     })
 
     it('should not check on buffer reload', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.checkBufferSwitch': false
-      })
+      helper.updateConfiguration('outline.checkBufferSwitch', false)
       await symbols.showOutline(1)
-      await helper.wait(50)
       await createBuffer()
       await helper.wait(50)
       let buf = await getOutlineBuffer()
       expect(buf).toBeDefined()
+    })
+
+    it('should sort by category', async () => {
+      let code = `
+class myClass {
+}
+fun1() {}
+`
+      await createBuffer(code)
+      await symbols.showOutline(1)
+      let buf = await getOutlineBuffer()
+      let lines = await buf.lines
+      expect(lines).toEqual([
+        'OUTLINE Category', '  c myClass 2', '  m fun1 4'
+      ])
     })
 
     it('should sort by position', async () => {
@@ -171,9 +191,7 @@ describe('symbols outline', () => {
   fun2() { }
   fun1() {}
 }`
-      workspace.configurations.updateUserConfig({
-        'outline.sortBy': 'position',
-      })
+      helper.updateConfiguration('outline.sortBy', 'position',)
       await createBuffer(code)
       await symbols.showOutline(1)
       let buf = await getOutlineBuffer()
@@ -188,9 +206,7 @@ describe('symbols outline', () => {
   fun2() {}
   fun1() {}
 }`
-      workspace.configurations.updateUserConfig({
-        'outline.sortBy': 'name',
-      })
+      helper.updateConfiguration('outline.sortBy', 'name')
       await createBuffer(code)
       await symbols.showOutline(1)
       let buf = await getOutlineBuffer()
@@ -201,9 +217,7 @@ describe('symbols outline', () => {
     })
 
     it('should change sort method', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.detailAsDescription': false
-      })
+      helper.updateConfiguration('outline.detailAsDescription', false)
       let code = `class detail {
   fun2() {}
   fun1() {}
@@ -222,9 +236,7 @@ describe('symbols outline', () => {
     })
 
     it('should show detail as description', async () => {
-      workspace.configurations.updateUserConfig({
-        'outline.detailAsDescription': true
-      })
+      helper.updateConfiguration('outline.detailAsDescription', true)
       let code = `class detail {
   fun2() {}
 }`
@@ -235,6 +247,18 @@ describe('symbols outline', () => {
       expect(lines.slice(1)).toEqual([
         '- c detail 1', '    m fun2 () 2'
       ])
+    })
+
+    it('should not showLineNumber', async () => {
+      helper.updateConfiguration('outline.showLineNumber', false)
+      let code = `class detail {
+  fun2() {}
+}`
+      await createBuffer(code)
+      await symbols.showOutline(1)
+      let buf = await getOutlineBuffer()
+      let lines = await buf.lines
+      expect(lines.slice(1)).toEqual(['- c detail', '    m fun2 ()'])
     })
   })
 
@@ -310,6 +334,8 @@ describe('symbols outline', () => {
       await nvim.command(`edit +setl\\ buftype=nofile t`)
       await workspace.document
       await symbols.showOutline(1)
+      let line = await helper.getCmdline()
+      expect(line).toMatch('Unable to show outline')
     })
 
     it('should not throw when provider does not exist', async () => {
@@ -386,6 +412,68 @@ describe('symbols outline', () => {
       await doc.synchronize()
       await symbols.showOutline(0)
       await helper.waitFor('getline', [1], 'OUTLINE vimlsp')
+    })
+  })
+
+  describe('autoPreview', () => {
+    it('should toggle auto preview by press p', async () => {
+      await createBuffer()
+      await symbols.showOutline(0)
+      await helper.waitFor('getline', [3], /fun1/)
+      await nvim.command('exe 2')
+      await nvim.input('p')
+      let winid = await helper.waitFloat()
+      expect(winid).toBeGreaterThan(1000)
+      await nvim.input('p')
+      await helper.waitValue(async () => {
+        let win = nvim.createWindow(winid)
+        let valid = await win.valid
+        return valid === false
+      }, true)
+    })
+
+    it('should close preview when move to line without node', async () => {
+      await createBuffer()
+      await symbols.showOutline(0)
+      await helper.waitFor('getline', [3], /fun1/)
+      await nvim.command('exe 2')
+      await nvim.input('p')
+      let winid = await helper.waitFloat()
+      await nvim.input('l')
+      await nvim.input('k')
+      await helper.waitValue(async () => {
+        let win = nvim.createWindow(winid)
+        let valid = await win.valid
+        return valid === false
+      }, true)
+    })
+
+    it('should show preview when move cursor back', async () => {
+      await createBuffer()
+      await symbols.showOutline(0)
+      await helper.waitFor('getline', [3], /fun1/)
+      await nvim.command('exe 2')
+      await nvim.input('p')
+      let winid = await helper.waitFloat()
+      await nvim.command('wincmd p')
+      await helper.waitValue(async () => {
+        let win = nvim.createWindow(winid)
+        let valid = await win.valid
+        return valid === false
+      }, true)
+      await nvim.command('wincmd p')
+      winid = await helper.waitFloat()
+      expect(winid).toBeGreaterThan(1000)
+    })
+
+    it('should enable auto preview by configuration', async () => {
+      helper.updateConfiguration('outline.autoPreview', true)
+      await createBuffer()
+      await symbols.showOutline(0)
+      await helper.waitFor('getline', [3], /fun1/)
+      await nvim.command('exe 2')
+      let winid = await helper.waitFloat()
+      expect(winid).toBeGreaterThan(1000)
     })
   })